Move energy cards to own folder
